Revert "Tribes Townsquare (#6495)"
This reverts commit beb8056f9557d2f47a9189b03c7c4aad3519334f.
