+++ app/assets/v2/css/users.css
@@ -117,6 +117,7 @@
 }
 
 .btn-search {
+  font-size: 1.3em;
   background: white;
   border-left: 0;
   color: #3E00FF;
