+++ app/assets/v2/js/pages/dashboard-hackathon.js
@@ -864,6 +864,122 @@
         hackathonSponsors: document.hackathonSponsors
       })
     });
+    Vue.component('project-directory', {
+      delimiters: [ '[[', ']]' ],
+      methods: {
+        openChat: function(handle) {
+          let vm = this;
+          const url = handle ? `${vm.chatURL}/hackathons/messages/@${handle}` : `${vm.chatURL}/`;
+
+          chatWindow = window.open(url, 'Loading', 'top=0,left=0,width=400,height=600,status=no,toolbar=no,location=no,menubar=no,titlebar=no');
+        },
+        fetchProjects: function(newPage) {
+          let vm = this;
+
+          vm.isLoading = true;
+          vm.noResults = false;
+
+          if (newPage) {
+            vm.projectsPage = newPage;
+          }
+          vm.params.page = vm.projectsPage;
+          vm.params.hackathon = hackathonId;
+
+          vm.params.filters = '';
+          if (vm.params.lfm) {
+            vm.params.filters = 'lfm';
+          }
+
+          if (vm.searchTerm) {
+            vm.params.search = vm.searchTerm;
+          } else {
+            delete vm.params['search'];
+          }
+
+          let searchParams = new URLSearchParams(vm.params);
+
+          let apiUrlProjects = `/api/v0.1/projects_fetch/?${searchParams.toString()}`;
+
+          var getProjects = fetchData(apiUrlProjects, 'GET');
+
+          $.when(getProjects).then(function(response) {
+            vm.hackathonProjects = [];
+            response.data.forEach(function(item) {
+              vm.hackathonProjects.push(item);
+            });
+
+            vm.projectsNumPages = response.num_pages;
+            vm.projectsHasNext = response.has_next;
+            vm.numProjects = response.count;
+            if (vm.projectsHasNext) {
+              vm.projectsPage = ++vm.projectsPage;
+
+            } else {
+              vm.projectsPage = 1;
+            }
+
+            if (vm.hackathonProjects.length) {
+              vm.noResults = false;
+            } else {
+              vm.noResults = true;
+            }
+            vm.isLoading = false;
+          });
+        },
+        searchProjects: function() {
+          let vm = this;
+
+          vm.hackathonProjects = [];
+
+          vm.fetchProjects(1);
+
+        }
+      },
+      data: () => ({
+        hackathonSponsors,
+        hackathonProjects,
+        projectsPage,
+        hackathonId,
+        projectsNumPages,
+        projectsHasNext,
+        numProjects,
+        media_url,
+        chatURL: document.chatURL,
+        lfm: false,
+        searchTerm: null,
+        bottom: false,
+        params: {},
+        isFunder: false,
+        showModal: false,
+        showFilters: true,
+        skills: document.keywords || [],
+        selectedSkills: [],
+        noResults: false,
+        isLoading: true,
+        hideFilterButton: false
+      }),
+      mounted() {
+        this.fetchProjects();
+        this.$watch('params', function(newVal, oldVal) {
+          this.searchProjects();
+        }, {
+          deep: true
+        });
+      },
+      created() {
+        // this.extractURLFilters();
+      },
+      beforeMount() {
+        window.addEventListener('scroll', () => {
+          this.bottom = this.bottomVisible();
+        }, false);
+      },
+      beforeDestroy() {
+        window.removeEventListener('scroll', () => {
+          this.bottom = this.bottomVisible();
+        });
+      }
+    });
     var app = new Vue({
       delimiters: [ '[[', ']]' ],
       el: '#dashboard-vue-app',
