+++ app/assets/v2/js/pages/tribe-edit.js
