+++ app/assets/v2/js/users.js
@@ -24,9 +24,7 @@ Vue.mixin({
         vm.usersPage = newPage;
       }
       vm.params.page = vm.usersPage;
-      if (hackathonId) {
-        vm.params.hackathon = hackathonId;
-      }
+      vm.params.hackathon = hackathonId;
       if (vm.searchTerm) {
         vm.params.search = vm.searchTerm;
       } else {
@@ -43,11 +41,6 @@ Vue.mixin({
         delete vm.params['organisation'];
         delete vm.params['skills'];
       }
-
-      if (vm.tribeFilter) {
-        vm.params.tribe = vm.tribeFilter;
-      }
-
 
       let searchParams = new URLSearchParams(vm.params);
 
@@ -253,7 +246,7 @@ Vue.mixin({
           --user.follower_count;
         }
 
-        event.target.classList.toggle('btn-outline-green');
+        event.target.classList.toggle('btn-gc-pink');
         event.target.classList.toggle('btn-gc-blue');
       }).fail(function(error) {
         event.target.disabled = false;
@@ -267,55 +260,39 @@ Vue = Vue.extend({
 
 Vue.component('user-directory', {
   delimiters: [ '[[', ']]' ],
-  props: [ 'tribe', 'is_my_org' ],
-  data: function() {
-    return {
-      orgOwner: this.is_my_org || false,
-      userFilter: {
-        options: [
-          { text: 'All', value: 'all' },
-          { text: 'Tribe Owners', value: 'owners' },
-          { text: 'Tribe Members', value: 'members' },
-          { text: 'Tribe Hackers', value: 'hackers' }
-        ]
-      },
-      tribeFilter: this.tribe || '',
-      users,
-      usersPage,
-      hackathonId,
-      usersNumPages,
-      usersHasNext,
-      numUsers,
-      media_url,
-      chatURL: document.chatURL || 'https://chat.gitcoin.co/',
-      searchTerm: null,
-      bottom: false,
-      params: {
-        'user_filter': 'all'
-      },
-      funderBounties: [],
-      currentBounty: undefined,
-      contributorInvite: undefined,
-      isFunder: false,
-      bountySelected: null,
-      userSelected: [],
-      showModal: false,
-      showFilters: !document.getElementById('explore_tribes'),
-      skills: document.keywords,
-      selectedSkills: [],
-      noResults: false,
-      isLoading: true,
-      gitcoinIssueUrl: '',
-      issueDetails: undefined,
-      errorIssueDetails: undefined,
-      showBanner: undefined,
-      persona: undefined,
-      hideFilterButton: !!document.getElementById('explore_tribes')
-    };
-  },
+  data: () => ({
+    users,
+    usersPage,
+    hackathonId,
+    usersNumPages,
+    usersHasNext,
+    numUsers,
+    media_url,
+    chatURL: document.chatURL || 'https://chat.gitcoin.co/',
+    searchTerm: null,
+    bottom: false,
+    params: {},
+    funderBounties: [],
+    currentBounty: undefined,
+    contributorInvite: undefined,
+    isFunder: false,
+    bountySelected: null,
+    userSelected: [],
+    showModal: false,
+    showFilters: !document.getElementById('explore_tribes'),
+    skills: document.keywords,
+    selectedSkills: [],
+    noResults: false,
+    isLoading: true,
+    gitcoinIssueUrl: '',
+    issueDetails: undefined,
+    errorIssueDetails: undefined,
+    showBanner: undefined,
+    persona: undefined,
+    hideFilterButton: !!document.getElementById('explore_tribes')
+  }),
   mounted() {
     this.fetchUsers();
-    this.tribeFilter = this.tribe;
     this.$watch('params', function(newVal, oldVal) {
       this.searchUsers();
     }, {
