+++ app/dashboard/models.py
@@ -66,7 +66,6 @@ from marketing.mails import featured_funded_bounty, fund_request_email, start_wo
 from marketing.models import LeaderboardRank
 from rest_framework import serializers
 from web3 import Web3
-from bounty_requests.models import BountyRequest
 
 from .notifications import maybe_market_to_github, maybe_market_to_slack, maybe_market_to_user_slack
 from .signals import m2m_changed_interested
@@ -75,7 +74,6 @@ logger = logging.getLogger(__name__)
 
 
 CROSS_CHAIN_STANDARD_BOUNTIES_OFFSET = 100000000
-
 
 class BountyQuerySet(models.QuerySet):
     """Handle the manager queryset for Bounties."""
@@ -2103,7 +2101,6 @@ class Activity(SuperModel):
         ('consolidated_leaderboard_rank', 'Consolidated Leaderboard Rank'),
         ('consolidated_mini_clr_payout', 'Consolidated CLR Payout'),
         ('hackathon_registration', 'Hackathon Registration'),
-        ('new_hackathon_project', 'New Hackathon Project'),
         ('flagged_grant', 'Flagged Grant'),
     ]
 
@@ -2153,12 +2150,6 @@ class Activity(SuperModel):
     hackathonevent = models.ForeignKey(
         'dashboard.HackathonEvent',
         related_name='activities',
-        on_delete=models.CASCADE,
-        blank=True, null=True
-    )
-    project = models.ForeignKey(
-        'dashboard.HackathonProject',
-        related_name='hackathon_projects',
         on_delete=models.CASCADE,
         blank=True, null=True
     )
@@ -2655,13 +2646,6 @@ class Profile(SuperModel):
     as_representation = JSONField(default=dict, blank=True)
     tribe_priority = models.TextField(default='', blank=True, help_text=_('HTML rich description for what tribe priorities.'))
 
-    tribes_cover_image = models.ImageField(
-        upload_to=get_upload_filename,
-        null=True,
-        blank=True,
-        help_text=_('The Tribes Cover image.'),
-    )
-
     is_org = models.BooleanField(
         default=True,
         help_text='Is this profile an org?',
@@ -2676,13 +2660,6 @@ class Profile(SuperModel):
 
     objects = ProfileManager()
     objects_full = ProfileQuerySet.as_manager()
-
-    @property
-    def suggested_bounties(self):
-        suggested_bounties = BountyRequest.objects.filter(tribe=self, status='o').order_by('created_on')
-
-        return suggested_bounties if suggested_bounties else []
-
     @property
     def subscribed_threads(self):
         tips = Tip.objects.filter(Q(pk__in=self.received_tips.all()) | Q(pk__in=self.sent_tips.all())).filter(comments_priv__icontains="activity:").all()
