+++ app/dashboard/router.py
@@ -29,8 +29,8 @@ from rest_framework import routers, serializers, viewsets
 from retail.helpers import get_ip
 
 from .models import (
-    Activity, Bounty, BountyFulfillment, BountyInvites, BountyRequest, HackathonEvent, HackathonProject, Interest, Profile,
-    ProfileSerializer, SearchHistory, TribeMember,
+    Activity, Bounty, BountyFulfillment, BountyInvites, HackathonEvent, HackathonProject, Interest, Profile,
+    ProfileSerializer, SearchHistory,
 )
 
 logger = logging.getLogger(__name__)
@@ -210,8 +210,7 @@ class BountySerializerSlim(BountySerializer):
             'pk', 'url', 'title', 'experience_level', 'status', 'fulfillment_accepted_on', 'event',
             'fulfillment_started_on', 'fulfillment_submitted_on', 'canceled_on', 'web3_created', 'bounty_owner_address',
             'avatar_url', 'network', 'standard_bounties_id', 'github_org_name', 'interested', 'token_name', 'value_in_usdt',
-            'keywords', 'value_in_token', 'project_type', 'is_open', 'expires_date', 'latest_activity', 'token_address',
-            'bounty_categories'
+            'keywords', 'value_in_token', 'project_type', 'is_open', 'expires_date', 'latest_activity', 'token_address'
         )
 
 
@@ -495,52 +494,6 @@ class BountyViewSet(viewsets.ModelViewSet):
         queryset = queryset.distinct()
 
         return queryset
-
-
-class TribesTeamSerializer(serializers.ModelSerializer):
-
-    user_is_following = serializers.SerializerMethodField(method_name='user_following')
-    followers_count = serializers.SerializerMethodField(method_name='follow_count')
-
-    def follow_count(self, instance):
-        return TribeMember.objects.filter(org=instance).exclude(status='rejected').exclude(profile__user=None).count()
-
-    def user_following(self, instance):
-        request = self.context.get('request')
-        user_profile = request.user.profile if request and request.user and hasattr(request.user, 'profile') else None
-        if user_profile:
-            return len(user_profile.tribe_members.filter(org__handle=instance.handle.lower())) > 0
-
-    class Meta:
-        model = Profile
-        fields = ('name', 'handle', 'avatar_url', 'followers_count', 'user_is_following')
-
-
-class BountyRequestSerializer(serializers.ModelSerializer):
-
-    requested_by = TribesTeamSerializer()
-
-    class Meta:
-        model = BountyRequest
-        fields = ('id', 'token_name', 'comment', 'github_url', 'title', 'requested_by', 'status')
-
-
-class TribesSerializer(serializers.ModelSerializer):
-    """Handle serializing the Profile object."""
-    active_bounties = BountySerializer(many=True)
-    team_or_none_if_timeout = TribesTeamSerializer(many=True, read_only=True)
-    suggested_bounties = BountyRequestSerializer(many=True)
-    tribes_cover_image = serializers.ImageField(allow_empty_file=True)
-
-    def __init__(self, *args, **kwargs):
-        super(TribesSerializer, self).__init__(*args, **kwargs)
-        # We pass the "upper serializer" context to the "nested one"
-        self.fields['team_or_none_if_timeout'].context.update(self.context)
-
-    class Meta:
-        model = Profile
-        """Define the profile serializer metadata."""
-        fields = ('profile_wallpaper', 'tribes_cover_image', 'name', 'linkedin_url', 'team_or_none_if_timeout', 'suggested_bounties', 'active_bounties', 'handle', 'tribe_description', 'avatar_url', 'follower_count', 'following_count', 'data', 'tribe_priority')
 
 
 # Routers provide an easy way of automatically determining the URL conf.
