+++ app/dashboard/templates/profiles/rank_tooltip.html
@@ -29,11 +29,9 @@
     "></i>
   </div>
 
-  {% if not skip_style %}
   <style>
     .tooltip-inner {
       max-width: 100% !important;
       text-align: left;
     }
   </style>
-  {% endif %}
