+++ app/dashboard/templates/profiles/tribe/suggest_bounty.html
@@ -173,3 +173,87 @@
     {% endif %}
   </div>
 </div>
+
+<style>
+  #post-to-ts {
+    top: 15px;
+  }
+
+  .priorities {
+    background-color: #1E2E42;
+  }
+
+  .text-grey {
+    color: rgba(0,0,0,0.50);
+  }
+
+  .text-white-70 {
+    color: rgba(255,255,255,0.70);
+  }
+
+  #priority-text p {
+    color: rgba(255,255,255,0.70) !important;
+  }
+
+  .suggested_bounties__box {
+    background-color: #EFF3F6;
+  }
+
+  .update-priority-submit {
+    background-color: #F8F8F8;
+  }
+
+  .update-priority {
+    background-color: white;
+  }
+
+  .update-priority img,
+  .priorities img {
+    max-width: 4rem;
+    height: auto;
+  }
+
+  .update-priority-submit .form__label {
+    font-size: 11px;
+    margin-top: 0.8rem;
+    color: #6587AE;
+  }
+
+  .update-priority-submit .form__label::before {
+    width: 15px;
+    height: 15px;
+    margin-right: 5px
+  }
+
+  .priority-container .ql-toolbar.ql-snow,
+  .priority-container .ql-container.ql-snow {
+    border-color: #EFF3F6;
+    border-left-color: white;
+    border-right-color: white;
+    border-bottom: 0px;
+  }
+
+  .priority-container .ql-container {
+    min-height: 5rem;
+  }
+
+  .suggested_bounties h2 {
+    color: rgba(0,0,0,0.70);
+  }
+
+  .suggest-bounty-form,
+  .suggested_bounty {
+    background: white;
+  }
+
+  @media (max-width: 767.98px) {
+    .priority-container .ql-container {
+      min-height: 4rem;
+    }
+
+    .priority-container .ql-toolbar.ql-snow,
+    .priority-container .ql-container.ql-snow {
+      border-color: #EFF3F6;
+    }
+  }
+</style>
