+++ app/dashboard/templates/shared/status_tooltip.html
@@ -22,7 +22,6 @@
     <li class="mb-2"><strong>{% trans "Work Started" %}:</strong> {% trans "Work has been started by at least one person." %}</li>
     <li class="mb-2"><strong>{% trans "Work Submitted" %}:</strong> {% trans "Completed work has been submitted by for review." %}</li>
     <li class="mb-2"><strong>{% trans "Work Done" %}:</strong> {% trans "The submitted project has been accepted and the funds have been paid." %}</li>
-    <li class="mb-2"><strong>{% trans "Reserved" %}:</strong> {% trans "The project has been reserved for an specific hunter." %}</li>
   </ol>
 
   <p class="font-smaller-1 font-weight-semibold">Possible Status</p>
