+++ app/dashboard/views.py
@@ -110,7 +110,7 @@ from .notifications import (
     maybe_market_tip_to_email, maybe_market_tip_to_github, maybe_market_tip_to_slack, maybe_market_to_email,
     maybe_market_to_github, maybe_market_to_slack, maybe_market_to_user_slack,
 )
-from .router import HackathonEventSerializer, HackathonProjectSerializer, TribesSerializer, TribesTeamSerializer
+from .router import HackathonEventSerializer, HackathonProjectSerializer
 from .utils import (
     apply_new_bounty_deadline, get_bounty, get_bounty_id, get_context, get_custom_avatars, get_unrated_bounties_count,
     get_web3, has_tx_mined, is_valid_eth_address, re_market_bounty, record_user_action_on_interest,
@@ -915,31 +915,23 @@ def projects_fetch(request):
     page = clean(request.GET.get('page', 1))
     hackathon_id = clean(request.GET.get('hackathon', ''))
 
-    if hackathon_id:
-        try:
-            hackathon_event = HackathonEvent.objects.get(id=hackathon_id)
-        except HackathonEvent.DoesNotExist:
-            hackathon_event = HackathonEvent.objects.last()
+    try:
+        hackathon_event = HackathonEvent.objects.get(id=hackathon_id)
+    except HackathonEvent.DoesNotExist:
+        hackathon_event = HackathonEvent.objects.last()
 
-        projects = HackathonProject.objects.filter(hackathon=hackathon_event).exclude(
-            status='invalid').prefetch_related('profiles', 'bounty').order_by(order_by, 'id')
+    projects = HackathonProject.objects.filter(hackathon=hackathon_event).exclude(status='invalid').prefetch_related('profiles', 'bounty').order_by(order_by, 'id')
 
-        if sponsor:
-            projects = projects.filter(
-                Q(bounty__github_url__icontains=sponsor)
-            )
-    elif sponsor:
-        sponsor_profile = Profile.objects.get(handle=sponsor)
-        if sponsor_profile:
-            projects = HackathonProject.objects.filter(hackathon__sponsor_profiles__in=[sponsor_profile]).exclude(
-                status='invalid').prefetch_related('profiles', 'bounty').order_by(order_by, 'id')
-        else:
-            projects = []
     if q:
         projects = projects.filter(
             Q(name__icontains=q) |
             Q(summary__icontains=q) |
             Q(profiles__handle__icontains=q)
+        )
+
+    if sponsor:
+        projects = projects.filter(
+            Q(bounty__github_url__icontains=sponsor)
         )
 
     if skills:
@@ -982,9 +974,7 @@ def users_fetch(request):
     leaderboard_rank = request.GET.get('leaderboard_rank', '').strip().split(',')
     rating = int(request.GET.get('rating', '0'))
     organisation = request.GET.get('organisation', '')
-    tribe = request.GET.get('tribe', '')
     hackathon_id = request.GET.get('hackathon', '')
-    user_filter = request.GET.get('user_filter', '')
 
     user_id = request.GET.get('user', None)
     if user_id:
@@ -1012,20 +1002,8 @@ def users_fetch(request):
     if q:
         profile_list = profile_list.filter(Q(handle__icontains=q) | Q(keywords__icontains=q))
 
-    if tribe:
-
-        profile_list = profile_list.filter(Q(follower__org__handle__in=[tribe]) | Q(organizations_fk__handle__in=[tribe]))
-
-        if user_filter and user_filter != 'all':
-            if user_filter == 'owners':
-                profile_list = profile_list.filter(Q(organizations_fk__handle__in=[tribe]))
-            elif user_filter == 'members':
-                profile_list = profile_list.exclude(Q(organizations_fk__handle__in=[tribe]))
-            elif user_filter == 'hackers':
-                profile_list = profile_list.filter(Q(fulfilled__bounty__github_url__icontains=tribe) | Q(project_profiles__hackathon__sponsor_profiles__handle__in=[tribe]) | Q(hackathon_registration__hackathon__sponsor_profiles__handle__in=[tribe]))
-
-
     show_banner = None
+
     if persona:
         if persona == 'funder':
             profile_list = profile_list.filter(dominant_persona='funder')
@@ -1077,11 +1055,7 @@ def users_fetch(request):
         this_page = Profile.objects_full.filter(pk__in=[ele.pk for ele in this_page]).order_by('-follower_count', 'id')
 
     else:
-        try:
-            profile_list = profile_list.order_by(order_by, '-earnings_count', 'id')
-        except profile_list.FieldError:
-            profile_list = profile_list.order_by('-earnings_count', 'id')
-
+        profile_list = profile_list.order_by(order_by, '-earnings_count', 'id')
         profile_list = profile_list.values_list('pk', flat=True)
 
         all_pages = Paginator(profile_list, limit)
@@ -2795,7 +2769,7 @@ def profile(request, handle, tab=None):
         context = {
             'hidden': True,
             'ratings': range(0,5),
-            'followers': TribeMember.objects.filter(org=request.user.profile) if request.user.is_authenticated and hasattr(request.user, 'profile') else [],
+            'followers': TribeMember.objects.filter(org=request.user.profile),
             'profile': {
                 'handle': handle,
                 'avatar_url': f"/dynamic/avatar/Self",
@@ -2836,29 +2810,18 @@ def profile(request, handle, tab=None):
     context['tab'] = tab
     context['show_activity'] = request.GET.get('p', False) != False
     context['is_my_org'] = request.user.is_authenticated and any([handle.lower() == org.lower() for org in request.user.profile.organizations ])
-    if request.user.is_authenticated and hasattr(request.user, 'profile'):
-        context['is_on_tribe'] = request.user.profile.tribe_members.filter(org__handle=handle.lower()).exists()
-    else:
-        context['is_on_tribe'] = False
+    context['is_on_tribe'] = False
+    if request.user.is_authenticated:
+        context['is_on_tribe'] = request.user.profile.tribe_members.filter(org__handle=handle.lower())
     context['ratings'] = range(0,5)
     context['feedbacks_sent'] = [fb.pk for fb in profile.feedbacks_sent.all() if fb.visible_to(request.user)]
     context['feedbacks_got'] = [fb.pk for fb in profile.feedbacks_got.all() if fb.visible_to(request.user)]
     context['all_feedbacks'] = context['feedbacks_got'] + context['feedbacks_sent']
     context['tags'] = [('#announce','bullhorn'), ('#mentor','terminal'), ('#jobs','code'), ('#help','laptop-code'), ('#other','briefcase'), ]
-    context['followers'] = TribeMember.objects.filter(org=request.user.profile) if request.user.is_authenticated and hasattr(request.user, 'profile') else []
-    context['following'] = TribeMember.objects.filter(profile=request.user.profile) if request.user.is_authenticated and hasattr(request.user, 'profile') else []
+    context['followers'] = TribeMember.objects.filter(org=request.user.profile)
+    context['following'] = TribeMember.objects.filter(profile=request.user.profile)
     context['foltab'] = request.GET.get('sub', 'followers')
 
-    active_tab = 0
-    if tab == "townsquare":
-        active_tab = 0
-    elif tab == "projects":
-        active_tab = 1
-    elif tab == "people":
-        active_tab = 2
-    elif tab == "bounties":
-        active_tab = 3
-    context['active_panel'] = active_tab
     tab = get_profile_tab(request, profile, tab, context)
     if type(tab) == dict:
         context.update(tab)
@@ -2868,16 +2831,6 @@ def profile(request, handle, tab=None):
     # record profile view
     if request.user.is_authenticated and not context['is_my_profile']:
         ProfileView.objects.create(target=profile, viewer=request.user.profile)
-
-    if profile.is_org and profile.handle.lower() in ['gitcoinco']:
-
-        context['currentProfile'] = TribesSerializer(profile, context={'request': request}).data
-        context['target'] = f'/activity?what=tribe:{profile.handle}'
-        context['is_on_tribe'] = json.dumps(context['is_on_tribe'])
-        context['is_my_org'] = json.dumps(context['is_my_org'])
-        context['profile_handle'] = profile.handle
-
-        return TemplateResponse(request, 'profiles/tribes-vue.html', context, status=status)
 
     return TemplateResponse(request, 'profiles/profile.html', context, status=status)
 
@@ -4546,18 +4499,13 @@ def join_tribe(request, handle):
     if request.user.is_authenticated:
         profile = request.user.profile if hasattr(request.user, 'profile') else None
         try:
-            try:
-                TribeMember.objects.get(profile=profile, org__handle=handle.lower()).delete()
-            except TribeMember.MultipleObjectsReturned:
-                TribeMember.objects.filter(profile=profile, org__handle=handle.lower()).delete()
-
+            TribeMember.objects.get(profile=profile, org__handle=handle.lower()).delete()
             return JsonResponse(
-                {
-                    'success': True,
-                    'is_member': False,
-                },
-                status=200
-            )
+            {
+                'success': True,
+                'is_member': False,
+            },
+            status=200)
         except TribeMember.DoesNotExist:
             kwargs = {
                 'org': Profile.objects.filter(handle=handle.lower()).first(),
@@ -4576,9 +4524,7 @@ def join_tribe(request, handle):
             )
     else:
         return JsonResponse(
-            {
-                'error': _('You must be authenticated via github to use this feature!')
-            },
+            { 'error': _('You must be authenticated via github to use this feature!') },
             status=401
         )
 
@@ -4664,13 +4610,6 @@ def save_tribe(request,handle):
             tribe = Profile.objects.filter(handle=handle.lower()).first()
             tribe.tribe_description = tribe_description
             tribe.save()
-
-        if request.FILES.get('cover_image'):
-            cover_image = request.FILES.get('cover_image', None)
-            if cover_image:
-                tribe = Profile.objects.filter(handle=handle.lower()).first()
-                tribe.tribes_cover_image = cover_image
-                tribe.save()
 
         if request.POST.get('tribe_priority'):
