+++ app/retail/templates/shared/activity.html
@@ -15,8 +15,7 @@
     along with this program. If not, see <http://www.gnu.org/licenses/>.
 
 {% endcomment %}
-
-{% load humanize i18n static grants_extra %}
+{% load humanize i18n static %}
 <div class="box d-block m-0 mb-3 activity {{ row.activity_type }} infinite-item px-sm-3 py-4" data-pk="{{row.pk}}">
 
   <div class="activity_main row">
@@ -50,9 +49,7 @@
         {% endif %}
         {% if row.profile.data.type == 'Organization'%}
           {% if row.profile.handle not in my_tribes|join:"," %}
-            <div style="text-align:center">
-              <a href=# class="align-middle btn btn-gc-blue btn-sm mt-2 font-smaller-7 font-weight-bold" onclick="joinTribeDirect(this);" data-jointribe='{{row.profile.handle}}'>follow</a>
-            </div>
+            <div style="text-align:center"><a href=# class="align-middle btn btn-gc-blue btn-sm mt-2 font-smaller-7 font-weight-bold" onclick="joinTribeDirect(this);" data-jointribe='{{row.profile.handle}}'>Join Tribe</a></div>
           {% endif %}
         {% endif %}
       </div>
@@ -363,35 +360,12 @@
         {% endif %}
       </div>
     </div>
-    <div class="d-none d-sm-inline col-sm-2 pl-0 position-relative">
-      {% if row.activity_type == 'new_bounty' %}
-        <div class="bg-gc-blue card-ribbon px-2 text-center">
-          <i style="font-size: 14px" class="text-white fas fa-tools"></i>
-          <span class="text-white font-weight-bold font-smaller-2 text-capitalize">New bounty</span>
-        </div>
-      {% elif row.activity_type == 'new_kudos' %}
-        <div class="bg-light-blue card-ribbon px-2 text-center">
-          <img height="15" src="{% static 'v2/images/top-bar/kudos_symbol_dark.svg' %}" alt="">
-          <span class="text-white font-weight-bold font-smaller-2 text-capitalize">New kudos</span>
-        </div>
-      {% elif row.activity_type == 'new_hackathon_project' %}
-        <div class="bg-gc-green card-ribbon px-2 text-center">
-          <img height="15" src="{% static 'v2/images/Lightbulb.svg' %}" alt="">
-          <span class="text-white font-weight-bold font-smaller-2 text-capitalize">New Project</span>
-        </div>
-      {% elif row.activity_type == 'new_grant' %}
-        <div class="bg-gc-dark-blue card-ribbon px-2 text-center">
-          <img height="15" src="{% static 'v2/images/products/grants-logo.svg' %}" alt="">
-          <span class="text-white font-weight-bold font-smaller-2 text-capitalize">New Grant</span>
-        </div>
-      {% endif %}
     <div class="d-none d-sm-inline col-sm-4 pl-0">
       <div class="activity_time font-smaller-5 text-right">
         {% if not hide_date %}
           {{ row.created_human_time }}
         {% endif %}
       </div>
-    </div>
       <div class="mt-3 text-right pr-3 activity-img">
         {% if row.kudos %}
           <a href="{{row.kudos.external_url}}">
